Explicitly list all project page routes to App
